refactor: update component name and files
